@@ -1,61 +1,59 @@
 {
-	"name": "imboard",
-	"version": "1.0.0",
-	"description": "",
-	"main": "index.js",
-	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1",
-		"start": "node index.js",
-		"server": "nodemon index.js",
-		"client": "react-scripts start",
-		"dev": "concurrently \"npm run server\" \"npm run client\"",
-		"build": "react-scripts build",
-		"heroku-postbuild": "NPM_CONFIG_PRODUCTION=false npm install && npm run build"
-	},
-	"keywords": [],
-	"author": "",
-	"license": "MIT",
-	"dependencies": {
-		"@fortawesome/fontawesome-svg-core": "^1.2.32",
-		"@fortawesome/free-brands-svg-icons": "^5.15.1",
-		"@fortawesome/free-solid-svg-icons": "^5.15.1",
-		"@fortawesome/react-fontawesome": "^0.1.12",
-		"bootstrap": "^4.5.3",
-		"concurrently": "^5.3.0",
-		"config": "^3.3.2",
-		"dotenv": "^8.2.0",
-		"express": "^4.17.1",
-		"http-proxy-middleware": "^1.0.6",
-		"mongoose": "^5.10.13",
-		"node-sass": "^4.14.1",
-		"react": "^17.0.1",
-		"react-dom": "^17.0.1",
-		"react-redux": "^7.2.2",
-		"react-router-dom": "^5.2.0",
-		"react-scripts": "^4.0.0",
-		"reactstrap": "^8.7.1",
-		"redux": "^4.0.5",
-		"redux-thunk": "^2.3.0",
-		"socket.io": "^3.0.1",
-		"socket.io-client": "^3.0.1"
-	},
-	"devDependencies": {
-		"nodemon": "^2.0.6",
-		"sass-lint": "^1.13.1"
-	},
-	"eslintConfig": {
-		"extends": "react-app"
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
-	}
+  "name": "imboard",
+  "version": "1.0.0",
+  "description": "A website to play all your favorite card and board games.",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "npm install && npm run build && cross-env NODE_ENV=production node index.js",
+    "build": "react-scripts build",
+    "dev-client": "react-scripts start",
+    "dev-server": "cross-env NODE_ENV=development nodemon index.js",
+    "dev-start": "concurrently \"npm run dev-server\" \"npm run dev-client\""
+  },
+  "keywords": [],
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-brands-svg-icons": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@reduxjs/toolkit": "^1.8.3",
+    "bootstrap": "^5.2.0",
+    "concurrently": "^7.3.0",
+    "config": "^3.3.7",
+    "cross-env": "^7.0.3",
+    "dotenv": "^16.0.1",
+    "express": "^4.18.1",
+    "http-proxy-middleware": "^2.0.6",
+    "mongoose": "^6.4.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "reactstrap": "^9.1.2",
+    "sass": "^1.53.0",
+    "socket.io": "^4.5.1",
+    "socket.io-client": "^4.5.1"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.19"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
