@@ -1,70 +1,81 @@
-import React, { useEffect, useContext } from 'react';
-import { Route, Switch } from 'react-router-dom';
+import React, { useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { Routes, Route } from 'react-router-dom';
 import CoupHome from './components/CoupHome';
 import CoupJoin from './components/CoupJoin';
 import CoupLobby from './components/CoupLobby';
 import CoupRoom from './components/CoupRoom';
 import NoRoom from '../components/NoRoom';
 import NoPage from '../components/NoPage';
-import socket from '../socket-client';
-import { CoupContext } from './context/CoupState';
+import { setRoom, addPlayer, removePlayer } from './roomSlice';
+import {
+  setMyPlayer,
+  setCards,
+  addCards,
+  keepCards,
+  setIsEliminated,
+} from './myPlayerSlice';
+import { coupSocket as socket } from '../socketClient';
 
 function CoupApp() {
-	const { room, setRoom, setPlayer, addPlayer, removePlayer } = useContext(
-		CoupContext
-	);
+  const room = useSelector((state) => state.room);
+  const dispatch = useDispatch();
 
-	useEffect(() => {
-		socket.on('update room', room => {
-			setRoom(room);
-		});
-		socket.on('update player', player => {
-			setPlayer(player);
-		});
-		socket.on('add player', playerName => {
-			addPlayer(playerName);
-		});
-		socket.on('remove player', playerName => {
-			removePlayer(playerName);
-		});
+  useEffect(() => {
+    socket.on('set room', (room) => {
+      dispatch(setRoom(room));
+    });
+    socket.on('add player', (player) => {
+      dispatch(addPlayer(player));
+    });
+    socket.on('remove player', (playerId) => {
+      dispatch(removePlayer(playerId));
+    });
+    socket.on('set my player', (myPlayer) => {
+      dispatch(setMyPlayer(myPlayer));
+    });
+    socket.on('set my cards', (cards) => {
+      dispatch(setCards(cards));
+    });
+    socket.on('add my cards', (cards) => {
+      dispatch(addCards(cards));
+    });
+    socket.on('keep my cards', (cardIndexes) => {
+      dispatch(keepCards(cardIndexes));
+    });
+    socket.on('set my elimination status', (isEliminated) => {
+      dispatch(setIsEliminated(isEliminated));
+    });
+    socket.on('disconnect', () => {
+      dispatch(setRoom(null));
+      dispatch(setMyPlayer(null));
+    });
 
-		return () => {
-			socket.off('update room');
-			socket.off('update player');
-			socket.off('add player');
-			socket.off('remove player');
-		};
-	}, [setRoom, setPlayer, addPlayer, removePlayer]);
+    return () => {
+      socket.off('set room');
+      socket.off('add player');
+      socket.off('remove player');
+      socket.off('set my player');
+      socket.off('set my cards');
+      socket.off('add my cards');
+      socket.off('keep my cards');
+      socket.off('set my elimination status');
+      socket.off('disconnect');
+    };
+  }, [dispatch]);
 
-	return (
-		<Switch>
-			<Route
-				exact
-				path="/coup"
-				render={({ history }) => <CoupHome history={history} />}
-			/>
-			<Route
-				exact
-				path="/coup/join"
-				render={({ history }) => <CoupJoin history={history} />}
-			/>
-			<Route
-				exact
-				path="/coup/lobby"
-				render={({ history }) =>
-					room !== null ? <CoupLobby history={history} /> : <NoRoom />
-				}
-			/>
-			<Route
-				exact
-				path="/coup/room"
-				render={({ history }) =>
-					room === null ? <NoRoom /> : <CoupRoom history={history} />
-				}
-			/>
-			<Route component={NoPage} />
-		</Switch>
-	);
+  return (
+    <Routes>
+      <Route path="/" element={<CoupHome />} />
+      <Route path="/join" element={<CoupJoin />} />
+      <Route
+        path="/lobby"
+        element={room !== null ? <CoupLobby /> : <NoRoom />}
+      />
+      <Route path="/room" element={room !== null ? <CoupRoom /> : <NoRoom />} />
+      <Route element={NoPage} />
+    </Routes>
+  );
 }
 
 export default CoupApp;
