@@ -1,7 +1,7 @@
-const server = require('./server');
-const config = require('./config');
-require('./io-server');
+const server = require('./server/server');
+const config = require('./server/config');
+require('./server/socket/socketServer');
 
-const PORT = config.PORT || 5000;
-
+const PORT = config.PORT || 8000;
+console.log(config.NODE_ENV);
 server.listen(PORT, () => console.log(`Server listening on port ${PORT}...`));
